Add time dependence of zodiacal light background
The zodiacal light background depends both on the sky coordinates
and the time of the observation because one needs to know both the
nearly fixed geocentric ecliptic coordinates of the target and the
time-dependent geocentric ecliptic coordinates of the sun.

Initially I had been simply converting the target coordinate to
heliocentric ecliptic coordinates based on a misunderstanding of
the STIS zodiacal light table which is stated in "helioecliptic"
coordinates. Really what matters here is the difference of
ecliptic longitudes of the target and of the sun.